consistent line spacing between functions
