+++ src/components/ratings-reviews/reviews/MoreAndAdd.jsx
@@ -11,18 +11,27 @@ export default function MoreAndAddReview() {
     shownReviews,
     lastReviewIdx,
   } = useRAndRContext();
-  const handleMoreClick = () => setCount(count + 2);
-  const handleAddClick = () => {
+
+  function handleMoreClick() {
+    setCount(count + 2);
+  }
+
+  function handleAddClick() {
     document.body.style.overflow = 'hidden';
     setShowWriteReview(true);
-  };
-  const handleCollapse = () => setCount(2);
+  }
+
+  function handleCollapse() {
+    setCount(2);
+  }
+
   useEffect(() => {
     if (count !== 2 && lastReviewIdx) {
       const lastReview = document.getElementById(`review-${shownReviews[lastReviewIdx].key}`);
       lastReview.scrollIntoView();
     }
   }, [count]);
+
   return (
     <div className="reviews-more-add">
       {showAdd && <button className="underline-button larger-text" type="button" onClick={handleMoreClick}>Show more</button>}
