+++ src/components/ratings-reviews/reviews/reviews-list/ReviewTile.jsx
@@ -3,33 +3,17 @@ import axios from 'axios';
 import { format, parseISO } from 'date-fns';
 import { StyledRatingStars } from '../../../../styled-lib';
 import { useRAndRContext } from '../../../../context/RAndRContext';
+import HighlightText from './HighlightText';
 
-function HighlightText({ text, highlight, bold = false }) {
-  if (highlight === '' || highlight === null) {
-    return (
-      <span style={bold ? { fontWeight: 'bold' } : {}}>{text}</span>
-    );
-  }
-  const regex = new RegExp(`(${highlight})`, 'gi');
-  const parts = text.split(regex);
-  return (
-    <span style={bold ? { fontWeight: 'bold' } : {}}>
-      {parts.filter((part) => part).map((part, index) => {
-        if (part.toLowerCase() === highlight.toLowerCase()) {
-          return <mark key={index}>{part}</mark>;
-        }
-        return <span key={index}>{part}</span>;
-      })}
-    </span>
-  );
-}
 export default function ReviewTile({ review, hidden, search }) {
   const [readMore, setReadMore] = useState(review.body.length > 250);
   const [showModal, setShowModal] = useState({ show: false, src: '' });
   const { reviewFeedback, setReviewFeedback } = useRAndRContext();
+
   function handleReadMoreClick() {
     setReadMore(!readMore);
   }
+
   function handlePhotoClick(event) {
     showModal.show ? (
       document.body.style.overflowX = 'hidden',
@@ -37,9 +21,11 @@ export default function ReviewTile({ review, hidden, search }) {
     ) : document.body.style.overflow = 'hidden';
     setShowModal({ show: !showModal.show, src: event.target.src });
   }
+
   function showPhotos() {
     return review.photos.map((item) => <img style={{ cursor: 'pointer', height: '100px' }} loading="lazy" key={item.id} src={item.url} alt={`review-${item.id}`} onClick={handlePhotoClick} />);
   }
+
   function handleHelpfulClick(e) {
     e.target.classList.add('clicked-link-button');
     e.target.removeAttribute('onClick');
@@ -55,6 +41,7 @@ export default function ReviewTile({ review, hidden, search }) {
         console.log('Error trying to mark review as helpful:', err);
       });
   }
+
   function handleReportClick(e) {
     e.target.classList.add('clicked-link-button');
     setReviewFeedback({
@@ -70,6 +57,7 @@ export default function ReviewTile({ review, hidden, search }) {
       })
       .catch((err) => console.log('Error trying to report review:', err));
   }
+
   function whichButton(name) {
     if (name === 'helpful') {
       return reviewFeedback.helpful.includes(review.review_id)
@@ -83,6 +71,7 @@ export default function ReviewTile({ review, hidden, search }) {
     }
     return <div />;
   }
+
   return (
     <>
       <div id={`review-${review.review_id}`} className="review-tile" hidden={hidden}>
